+++ nx.json
@@ -13,7 +13,7 @@
       "{projectRoot}/src/**/*",
       "{projectRoot}/tsconfig.json",
       "{projectRoot}/package.json",
-      "{projectRoot}/rslib.config.*"
+      "{projectRoot}/rstack.config.*"
     ],
     "dev": ["default"],
     "build": ["default"],
