+++ packages/core/package.json
@@ -58,8 +58,8 @@
     "index.html"
   ],
   "scripts": {
-    "build": "node ./scripts/generateJsonSchema.mjs && rslib build",
-    "dev": "rslib build -w",
+    "build": "node ./scripts/generateJsonSchema.mjs && rs lib",
+    "dev": "rs lib -w",
     "reset": "rimraf ./**/node_modules"
   },
   "dependencies": {
@@ -106,7 +106,6 @@
     "@microsoft/api-extractor": "^7.59.0",
     "@rsbuild/plugin-sass": "~2.0.1",
     "@rsbuild/plugin-svgr": "^2.0.5",
-    "@rslib/core": "1.0.0-rc.0",
     "@rspress/config": "workspace:*",
     "@types/body-scroll-lock": "^3.1.2",
     "@types/hast": "^3.0.5",
