+++ packages/core/src/env.d.ts
@@ -1,5 +1,3 @@
-/// <reference types='@rslib/core/types' />
-
 interface ImportMetaEnv {
   readonly ENABLE_LLMS_HINT: boolean;
   readonly ENABLE_LLMS_UI: boolean;
