+++ packages/core/src/node/env.d.ts
@@ -1,5 +1,3 @@
-/// <reference types='@rslib/core/types' />
-
 declare global {
   const REQUIRE_CACHE: NodeRequire['cache'];
 }
