+++ packages/create-rspress/package.json
@@ -23,8 +23,8 @@
     "bin.js"
   ],
   "scripts": {
-    "build": "node ./scripts/generate-theme-templates.mjs && rslib build",
-    "dev": "node ./scripts/generate-theme-templates.mjs && rslib build -w",
+    "build": "node ./scripts/generate-theme-templates.mjs && rs lib",
+    "dev": "node ./scripts/generate-theme-templates.mjs && rs lib -w",
     "start": "node ./dist/index.js"
   },
   "dependencies": {
@@ -32,7 +32,6 @@
   },
   "devDependencies": {
     "@microsoft/api-extractor": "^7.59.0",
-    "@rslib/core": "1.0.0-rc.0",
     "@types/node": "^22.8.1",
     "rsbuild-plugin-publint": "^1.0.0",
     "typescript": "^6.0.3"
