--- packages/create-rspress/rslib.config.ts
@@ -1,11 +1,0 @@
-import { defineConfig } from '@rslib/core';
-import { pluginPublint } from 'rsbuild-plugin-publint';
-
-export default defineConfig({
-  plugins: [pluginPublint()],
-  lib: [
-    {
-      syntax: 'es2023',
-    },
-  ],
-});
