+++ packages/create-rspress/rstack.config.ts
@@ -1,0 +1,7 @@
+import { define } from 'rstack';
+import { pluginPublint } from 'rsbuild-plugin-publint';
+
+define.lib({
+  plugins: [pluginPublint()],
+  syntax: 'es2023',
+});
