+++ packages/plugin-algolia/package.json
@@ -34,8 +34,8 @@
     "static"
   ],
   "scripts": {
-    "build": "rslib build",
-    "dev": "rslib build -w",
+    "build": "rs lib",
+    "dev": "rs lib -w",
     "reset": "rimraf ./**/node_modules"
   },
   "dependencies": {
@@ -45,7 +45,6 @@
   "devDependencies": {
     "@microsoft/api-extractor": "^7.59.0",
     "@rsbuild/plugin-react": "~2.1.0",
-    "@rslib/core": "1.0.0-rc.0",
     "@rspress/config": "workspace:*",
     "@types/node": "^22.8.1",
     "@types/react": "^19.2.18",
