+++ packages/plugin-algolia/src/env.d.ts
@@ -1,2 +1,1 @@
-/// <reference types='@rslib/core/types' />
 declare module '@docsearch/css';
