+++ packages/plugin-api-docgen/rstack.config.ts
@@ -1,18 +1,14 @@
 import { fileURLToPath } from 'node:url';
-import { defineConfig } from '@rslib/core';
+import { define } from 'rstack';
 import { pluginPublint } from 'rsbuild-plugin-publint';
 
 const typescriptPath = fileURLToPath(import.meta.resolve('@typescript/native'));
 
-export default defineConfig({
+define.lib({
   plugins: [pluginPublint()],
-  lib: [
-    {
-      dts: {
-        typescriptPath,
-        bundle: true,
-      },
-      syntax: 'es2023',
-    },
-  ],
+  dts: {
+    typescriptPath,
+    bundle: true,
+  },
+  syntax: 'es2023',
 });
