--- packages/plugin-client-redirects/src/env.d.ts
@@ -1,1 +1,0 @@
-/// <reference types='@rslib/core/types' />
