+++ packages/plugin-llms/package.json
@@ -28,8 +28,8 @@
     "static"
   ],
   "scripts": {
-    "build": "rslib build",
-    "dev": "rslib build -w",
+    "build": "rs lib",
+    "dev": "rs lib -w",
     "reset": "rimraf ./**/node_modules"
   },
   "dependencies": {
@@ -43,7 +43,6 @@
     "@microsoft/api-extractor": "^7.59.0",
     "@rsbuild/core": "^2.1.13",
     "@rsbuild/plugin-react": "~2.1.0",
-    "@rslib/core": "1.0.0-rc.0",
     "@rspress/config": "workspace:*",
     "@types/hast": "^3.0.5",
     "@types/mdast": "^4.0.4",
