+++ packages/plugin-playground/package.json
@@ -27,8 +27,8 @@
     "static"
   ],
   "scripts": {
-    "build": "rslib build",
-    "dev": "rslib build -w",
+    "build": "rs lib",
+    "dev": "rs lib -w",
     "reset": "rimraf ./**/node_modules"
   },
   "dependencies": {
@@ -42,7 +42,6 @@
   "devDependencies": {
     "@babel/types": "^7.29.8",
     "@rsbuild/plugin-react": "~2.1.0",
-    "@rslib/core": "1.0.0-rc.0",
     "@rspress/plugin-playground": "workspace:*",
     "@types/babel__core": "^7.20.5",
     "@types/babel__standalone": "^7.1.9",
