+++ packages/plugin-playground/static/env.d.ts
@@ -1,5 +1,3 @@
-/// <reference types='@rslib/core/types' />
-
 declare module '_rspress_playground_imports' {
   const getImport: (name: string, getDefault?: boolean) => void;
   export default getImport;
