+++ packages/plugin-preview/rstack.config.ts
@@ -1,10 +1,10 @@
 import { fileURLToPath } from 'node:url';
-import { defineConfig } from '@rslib/core';
+import { define } from 'rstack';
 import { pluginPublint } from 'rsbuild-plugin-publint';
 
 const typescriptPath = fileURLToPath(import.meta.resolve('@typescript/native'));
 
-export default defineConfig({
+define.lib({
   plugins: [pluginPublint()],
   lib: [
     {
