+++ packages/plugin-preview/static/env.d.ts
@@ -1,5 +1,3 @@
-/// <reference types='@rslib/core/types' />
-
 declare module 'virtual-meta' {
   const demos: Record<string, []>;
   export { demos };
