+++ packages/plugin-rss/package.json
@@ -25,15 +25,14 @@
     "index.d.ts"
   ],
   "scripts": {
-    "build": "rslib build",
-    "dev": "rslib build -w",
+    "build": "rs lib",
+    "dev": "rs lib -w",
     "reset": "rimraf ./**/node_modules"
   },
   "dependencies": {
     "feed": "^5.2.1"
   },
   "devDependencies": {
-    "@rslib/core": "1.0.0-rc.0",
     "@types/node": "^22.8.1",
     "@types/react": "^19.2.18",
     "react": "^19.2.8",
