+++ packages/plugin-sitemap/package.json
@@ -24,13 +24,12 @@
     "static"
   ],
   "scripts": {
-    "build": "rslib build",
-    "dev": "rslib build -w",
+    "build": "rs lib",
+    "dev": "rs lib -w",
     "reset": "rimraf ./**/node_modules"
   },
   "devDependencies": {
     "@microsoft/api-extractor": "^7.59.0",
-    "@rslib/core": "1.0.0-rc.0",
     "@rspress/config": "workspace:*",
     "rsbuild-plugin-publint": "^1.0.0"
   },
