+++ packages/plugin-sitemap/rstack.config.ts
@@ -1,19 +1,15 @@
 import { fileURLToPath } from 'node:url';
-import { defineConfig } from '@rslib/core';
+import { define } from 'rstack';
 import { pluginPublint } from 'rsbuild-plugin-publint';
 
 const typescriptPath = fileURLToPath(import.meta.resolve('@typescript/native'));
 
-export default defineConfig({
+define.lib({
   plugins: [pluginPublint()],
-  lib: [
-    {
-      bundle: true,
-      syntax: 'es2023',
-      dts: {
-        typescriptPath,
-        bundle: true,
-      },
-    },
-  ],
+  bundle: true,
+  syntax: 'es2023',
+  dts: {
+    typescriptPath,
+    bundle: true,
+  },
 });
