+++ packages/shared/package.json
@@ -45,9 +45,9 @@
     "dist"
   ],
   "scripts": {
-    "build": "rslib build",
-    "build:watch": "rslib build -w",
-    "dev": "rslib build -w",
+    "build": "rs lib",
+    "build:watch": "rs lib -w",
+    "dev": "rs lib -w",
     "reset": "rimraf ./**/node_modules"
   },
   "dependencies": {
@@ -58,7 +58,6 @@
     "unified": "^11.0.5"
   },
   "devDependencies": {
-    "@rslib/core": "1.0.0-rc.0",
     "@rspress/config": "workspace:*",
     "@types/lodash-es": "^4.17.12",
     "@types/node": "^22.8.1",
