+++ packages/shared/rstack.config.ts
@@ -1,19 +1,15 @@
 import { fileURLToPath } from 'node:url';
-import { defineConfig } from '@rslib/core';
+import { define } from 'rstack';
 import { pluginPublint } from 'rsbuild-plugin-publint';
 
 const typescriptPath = fileURLToPath(import.meta.resolve('@typescript/native'));
 
-export default defineConfig({
-  lib: [
-    {
-      dts: {
-        typescriptPath,
-        bundle: true,
-      },
-      syntax: 'es2023',
-    },
-  ],
+define.lib({
+  dts: {
+    typescriptPath,
+    bundle: true,
+  },
+  syntax: 'es2023',
   source: {
     entry: {
       index: 'src/index.ts',
